backend/requirements.txt
uvicorn[standard]==0.21.1
python-multipart
requests
sqlalchemy==1.4.54
databases==0.6.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
openai-whisper 
torch     
pydub
PyJWT
ffmpeg
gTTS
aiofiles
openai>=0.27.0
chromadb>=0.4.3
sentence-transformers>=2.2.2
transformers>=4.30.0
weaviate-client>=4.18.3  
fastapi==0.109.2
starlette==0.36.3
httpx==0.28.1
pytest>=7.3.0
langchain>=0.3.27
langchain-openai>=0.1.0
langchain-community>=0.2.0
pydantic>=2.8.0
sqlalchemy>=1.4
httpx
alembic
langsmith        
autogen          
respx
pyttsx3
elevenlabs
neo4j
yt-dlp
ffmpeg-python
evaluate
datasets
rouge-score
peft
accelerate
prometheus-client